requirements.txt
fastapi
sqlalchemy
uvicorn
python-dotenv
mysql
pytest
httpx
pydantic
structlog
siegeapi
